+++ tower-http/src/services/fs/serve_dir/open_file.rs
@@ -348,8 +348,10 @@ fn append_slash_on_path(uri: Uri) -> Result<Uri, OpenFileOutput> {
         uri_builder.path_and_query("/")
     };
 
-    uri_builder.build().map_err(|err| {
-        tracing::error!(?err, "redirect uri failed to build");
+    uri_builder.build().map_err(|_err| {
+        #[cfg(feature = "tracing")]
+        tracing::error!(err = ?_err, "redirect uri failed to build");
+
         OpenFileOutput::InvalidRedirectUri
     })
 }